+++ pg_lakehouse/src/fdw/mod.rs
@@ -1,4 +1,3 @@
 pub mod format;
 pub mod handler;
-pub mod object_store;
 pub mod options;
