+++ pg_lakehouse/src/stores/mod.rs
@@ -1,4 +1,4 @@
-mod base;
+pub mod base;
 pub mod local;
 pub mod s3;
 pub mod utils;
