+++ pg_lakehouse/src/stores/s3.rs
@@ -29,6 +29,116 @@ pub(crate) struct S3Fdw {
     target_columns: Vec<Column>,
 }
 
+pub enum AmazonServerOption {
+    Endpoint,
+    Bucket,
+    Root,
+    Region,
+    AllowAnonymous,
+}
+
+impl AmazonServerOption {
+    pub fn as_str(&self) -> &str {
+        match self {
+            Self::Endpoint => "endpoint",
+            Self::Bucket => "bucket",
+            Self::Root => "root",
+            Self::Region => "region",
+            Self::AllowAnonymous => "allow_anonymous",
+        }
+    }
+
+    pub fn is_required(&self) -> bool {
+        match self {
+            Self::Endpoint => false,
+            Self::Bucket => true,
+            Self::Root => false,
+            Self::Region => false,
+            Self::AllowAnonymous => false,
+        }
+    }
+
+    pub fn iter() -> impl Iterator<Item = Self> {
+        [
+            Self::Endpoint,
+            Self::Bucket,
+            Self::Root,
+            Self::Region,
+            Self::AllowAnonymous,
+        ]
+        .into_iter()
+    }
+}
+
+pub enum AmazonUserMappingOption {
+    AccessKeyId,
+    SecretAccessKey,
+    SecurityToken,
+}
+
+impl AmazonUserMappingOption {
+    pub fn as_str(&self) -> &str {
+        match self {
+            Self::AccessKeyId => "access_key_id",
+            Self::SecretAccessKey => "secret_access_key",
+            Self::SecurityToken => "security_token",
+        }
+    }
+}
+
+impl TryFrom<ServerOptions> for S3 {
+    type Error = ContextError;
+
+    fn try_from(options: ServerOptions) -> Result<Self, Self::Error> {
+        let server_options = options.server_options();
+        let user_mapping_options = options.user_mapping_options();
+
+        let mut builder = S3::default();
+        let bucket = require_option(AmazonServerOption::Bucket.as_str(), server_options)?;
+        builder.bucket(bucket);
+
+        if let Some(root) = server_options.get(AmazonServerOption::Root.as_str()) {
+            builder.root(root);
+        }
+
+        if let Some(region) = server_options.get(AmazonServerOption::Region.as_str()) {
+            builder.region(region);
+        }
+
+        if let Some(access_key_id) =
+            user_mapping_options.get(AmazonUserMappingOption::AccessKeyId.as_str())
+        {
+            builder.access_key_id(access_key_id);
+        }
+
+        if let Some(secret_access_key) =
+            user_mapping_options.get(AmazonUserMappingOption::SecretAccessKey.as_str())
+        {
+            builder.secret_access_key(secret_access_key);
+        }
+
+        if let Some(security_token) =
+            user_mapping_options.get(AmazonUserMappingOption::SecurityToken.as_str())
+        {
+            builder.security_token(security_token);
+        }
+
+        if let Some(endpoint) = server_options.get(AmazonServerOption::Endpoint.as_str()) {
+            builder.endpoint(endpoint);
+        }
+
+        if let Some(allow_anonymous) =
+            server_options.get(AmazonServerOption::AllowAnonymous.as_str())
+        {
+            if allow_anonymous == "true" {
+                builder.allow_anonymous();
+            }
+        }
+
+        Ok(builder)
+    }
+}
+
 impl BaseFdw for S3Fdw {
     fn register_object_store(
         server_options: HashMap<String, String>,
@@ -136,6 +246,12 @@ impl ForeignDataWrapper<BaseFdwError> for S3Fdw {
             match oid {
                 FOREIGN_DATA_WRAPPER_RELATION_ID => {}
                 FOREIGN_SERVER_RELATION_ID => {
+                    let valid_options: Vec<String> = AmazonServerOption::iter()
+                        .map(|opt| opt.as_str().to_string())
+                        .collect();
+
+                    validate_options(opt_list.clone(), valid_options)?;
+
                     for opt in AmazonServerOption::iter() {
                         if opt.is_required() {
                             check_options_contain(&opt_list, opt.as_str())?;
@@ -143,6 +259,12 @@ impl ForeignDataWrapper<BaseFdwError> for S3Fdw {
                     }
                 }
                 FOREIGN_TABLE_RELATION_ID => {
+                    let valid_options: Vec<String> = TableOption::iter()
+                        .map(|opt| opt.as_str().to_string())
+                        .collect();
+
+                    validate_options(opt_list.clone(), valid_options)?;
+
                     for opt in TableOption::iter() {
                         if opt.is_required() {
                             check_options_contain(&opt_list, opt.as_str())?;
